python-mock-physical-twin/requirements.txt
PyYAML==6.0.2
paho-mqtt==1.6.1
requests==2.31.0
Flask==2.3.2
